+++ js_concurrent_module/taskpool/sequence_runner.cpp
@@ -166,7 +166,7 @@ napi_value SequenceRunner::Execute(napi_env env, napi_callback_info cbinfo)
     return promise;
 }
 
-void SequenceRunner::ExecuteTaskImmediately(uint64_t taskId, Priority priority)
+void SequenceRunner::ExecuteTaskImmediately(uint32_t taskId, Priority priority)
 {
     TaskManager::GetInstance().EnqueueTaskId(taskId, priority);
 }
